WIP: Vastly simplify image wallpaper code
The image wallpaper is a mess of 4 different fade functions, some not in
sync, and very complicated swapping images about.

This simplifies the whole thing by using a StackView, we just pop our
setup with a new configuration on and it handles deleting old objects
when the transition ends.

Code is 1/3 the size and should result in having fewer items about.

In addition with this patch changing the transition effect in future or
via config would be super trivial.