use a separate kind of package for splash screens
use a standalone package type for splashscreens retrocompatible with
lnf packages with only a splash in it, to be installed separately.
try to load one of them first, if fails go to the lnf route to
be compatible. splash screens will be in a new separed category again
in the KDE store, so newly installed ones will go in plasma/splashscreens
the format is unchanged